Refactor employee routes: move to new directory structure and update imports
